requirements_testing.txt
# requirements for running tests (on top of environment.yml/requirements.txt)
pytest!=4.6.0
pytest-cov
pytest-timeout
pytest-harvest
flake8
flake8-array-spacing
numpydoc
codespell
pydocstyle
check-manifest
twine
wheel
pooch
mne-bids
importlib_resources
pymatreader
black
isort
mypy
matplotlib
python-picard
joblib
scikit-learn
pooch